modernize the codebase
