+++ readme.md
@@ -1,10 +1,15 @@
-# download-tarball 
+# download-tarball
 
-Download a tarball (optionally gzipped) to a folder &amp; extract it in the process. Uses the wonderful &amp; super quick tar-fs &amp; gunzip-maybe libraries.
+Download a tarball (optionally gzipped) to a folder &amp; extract it in the process. Uses the wonderful &amp; super quick tar-fs &amp; gunzip-maybe srcraries.
 
 ## Installation
 
-Download node at [nodejs.org](http://nodejs.org) and install it, if you haven't already.
+This is a [Node.js](https://nodejs.org/) module available through the 
+[npm registry](https://www.npmjs.com/). It can be installed using the 
+[`npm`](https://docs.npmjs.com/getting-started/installing-npm-packages-locally)
+or 
+[`yarn`](https://yarnpkg.com/en/)
+command line tools.
 
 ```sh
 npm install download-tarball --save
@@ -65,30 +70,27 @@ npm test
 
 ## Dependencies
 
-- [got](https://github.com/sindresorhus/got): Simplified HTTP requests
-- [gunzip-maybe](https://github.com/mafintosh/gunzip-maybe): Transform stream that gunzips its input if it is gzipped and just echoes it if not
-- [object-assign](https://github.com/sindresorhus/object-assign): ES2015 `Object.assign()` ponyfill
-- [promisify-function](https://github.com/jcollado/promisify-function): Turn a callback-style function into a function that returns a promise
-- [pump](https://github.com/mafintosh/pump): pipe streams together and close all of them if one of them closes
-- [tar-fs](https://github.com/mafintosh/tar-fs): filesystem bindings for tar-stream
+- [gunzip-maybe](https://ghub.io/gunzip-maybe): Transform stream that gunzips its input if it is gzipped and just echoes it if not
+- [pump](https://ghub.io/pump): pipe streams together and close all of them if one of them closes
+- [tar-fs](https://ghub.io/tar-fs): filesystem bindings for tar-stream
 
 ## Dev Dependencies
 
-- [babel-cli](https://github.com/babel/babel/tree/master/packages): Babel command line.
-- [babel-core](https://github.com/babel/babel/tree/master/packages): Babel compiler core.
-- [babel-preset-es2015](https://github.com/babel/babel/tree/master/packages): Babel preset for all es2015 plugins.
-- [babel-tape-runner](https://github.com/wavded/babel-tape-runner): Babel + Tape for running your ES Next tests
-- [http-test-server](https://github.com/kesla/test-http-server): Create a simple http server for tests
-- [package-json-to-readme](https://github.com/zeke/package-json-to-readme): Generate a README.md from package.json contents
-- [tapava](https://github.com/kesla/tapava): the syntax of ava, run through tape
-- [tar-stream](https://github.com/mafintosh/tar-stream): tar-stream is a streaming tar parser and generator and nothing else. It is streams2 and operates purely using streams which means you can easily extract/parse tarballs without ever hitting the file system.
-- [then-fs](https://github.com/then/fs): promised FS
-- [tmp](https://github.com/raszi/node-tmp): Temporary file and directory creator
-- [xo](https://github.com/sindresorhus/xo): JavaScript happiness style linter ❤️
-
+- [@types/gunzip-maybe](https://ghub.io/@types/gunzip-maybe): TypeScript definitions for gunzip-maybe
+- [@types/jest](https://ghub.io/@types/jest): TypeScript definitions for Jest
+- [@types/node](https://ghub.io/@types/node): TypeScript definitions for Node.js
+- [@types/pump](https://ghub.io/@types/pump): TypeScript definitions for pump
+- [@types/tar-fs](https://ghub.io/@types/tar-fs): TypeScript definitions for tar-fs
+- [http-test-server](https://ghub.io/http-test-server): Create a simple http server for tests
+- [jest](https://ghub.io/jest): Delightful JavaScript Testing.
+- [package-json-to-readme](https://ghub.io/package-json-to-readme): Generate a README.md from package.json contents
+- [tar-stream](https://ghub.io/tar-stream): tar-stream is a streaming tar parser and generator and nothing else. It operates purely using streams which means you can easily extract/parse tarballs without ever hitting the file system.
+- [tmp](https://ghub.io/tmp): Temporary file and directory creator
+- [ts-jest](https://ghub.io/ts-jest): A Jest transformer with source map support that lets you use Jest to test projects written in TypeScript
+- [tsup](https://ghub.io/tsup): Bundle your TypeScript library with no config, powered by esbuild
+- [got](https://ghub.io/got): Human-friendly and powerful HTTP request library for Node.js
+- [typescript](https://ghub.io/typescript): TypeScript is a language for application scale JavaScript development
 
 ## License
 
 MIT
-
-_Generated by [package-json-to-readme](https://github.com/zeke/package-json-to-readme)_
